feat: add CRUD Endereco

+++ src/main/java/com/example/hotmartdesafio/repositories/EnderecoRepository.java
@@ -3,6 +3,6 @@ package com.example.hotmartdesafio.repositories;
 import com.example.hotmartdesafio.models.Endereco;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EnderecoRepository extends CrudRepository<Endereco, Integer> {
+public interface EnderecoRepository extends CrudRepository<Endereco, Long> {
 
 }
